Conflicts:
	src/de/_13ducks/spacebatz/client/Client.java

+++ src/de/_13ducks/spacebatz/client/Position.java
@@ -6,9 +6,9 @@ package de._13ducks.spacebatz.client;
  * @author Jojo
  */
 public class Position {
-    private int x, y;
+    private double x, y;
     
-    public Position(int x, int y) {
+    public Position(double x, double y) {
         this.x = x;
         this.y = y;
     }
@@ -16,28 +16,28 @@ public class Position {
     /**
      * @return the x
      */
-    public int getX() {
+    public double getX() {
         return x;
     }
 
     /**
      * @param x the x to set
      */
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
     /**
      * @return the y
      */
-    public int getY() {
+    public double getY() {
         return y;
     }
 
     /**
      * @param y the y to set
      */
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 }
